Fixes #57 (Transition Amount to BigDecimal) (#73)
* Transitions all java-ilp-core “amount” fields to be BigDecimal.
* Add unit tests for Quote model objects and builders.
* Add unit tests for negative amounts in Quote constructors
* Fix unit tests and Checkstyle errors.
* Update version to `0.7.0-SNAPSHOT` due to interface change
* Fix checkstyle location in pom file to align with gradle build.
* Misc Code Review Fixes
* Fix codacy errors.

+++ src/test/java/org/interledger/codecs/oer/ilqp/IlqpCodecTests.java
@@ -5,9 +5,11 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 import org.interledger.InterledgerAddress;
 import org.interledger.InterledgerPacket;
+
 import org.interledger.codecs.Codec;
 import org.interledger.codecs.CodecContext;
 import org.interledger.codecs.CodecContextFactory;
+
 import org.interledger.ilqp.LiquidityCurve;
 import org.interledger.ilqp.LiquidityPoint;
 import org.interledger.ilqp.QuoteByDestinationAmountRequest;
@@ -46,34 +48,34 @@ public class IlqpCodecTests {
    */
   @Parameters
   public static Object[] data() {
-    return new Object[] {
+    return new Object[]{
         QuoteBySourceAmountRequest.Builder.builder()
             .destinationAccount(InterledgerAddress.of("test1.foo"))
-            .sourceAmount(100)
+            .sourceAmount(BigInteger.valueOf(100L))
             .destinationHoldDuration(Duration.ofSeconds(30)).build(),
         QuoteBySourceAmountResponse.Builder.builder()
-            .destinationAmount(95)
+            .destinationAmount(BigInteger.valueOf(95L))
             .sourceHoldDuration(Duration.ofSeconds(30)).build(),
         QuoteByDestinationAmountRequest.Builder.builder()
             .destinationAccount(InterledgerAddress.of("test2.foo"))
-            .destinationAmount(100)
+            .destinationAmount(BigInteger.valueOf(100L))
             .destinationHoldDuration(Duration.ofSeconds(35)).build(),
         QuoteByDestinationAmountResponse.Builder.builder()
-            .sourceAmount(105)
+            .sourceAmount(BigInteger.valueOf(105L))
             .sourceHoldDuration(Duration.ofMinutes(1)).build(),
         QuoteLiquidityRequest.Builder.builder()
             .destinationAccount(InterledgerAddress.of("test3.foo"))
-        .destinationHoldDuration(Duration.ofMinutes(5)).build(),
+            .destinationHoldDuration(Duration.ofMinutes(5)).build(),
         QuoteLiquidityResponse.Builder.builder()
             .liquidityCurve(
                 LiquidityCurve.Builder
                     .builder()
-                      .liquidityPoint(LiquidityPoint.Builder.builder()
+                    .liquidityPoint(LiquidityPoint.Builder.builder()
                         .inputAmount(BigInteger.ZERO).outputAmount(BigInteger.ZERO).build())
-                      .liquidityPoint(LiquidityPoint.Builder.builder()
-                          .inputAmount(BigInteger.ONE).outputAmount(BigInteger.ONE).build())
-                      .liquidityPoint(LiquidityPoint.Builder.builder()
-                          .inputAmount(BigInteger.valueOf(5)).outputAmount(BigInteger.TEN).build())
+                    .liquidityPoint(LiquidityPoint.Builder.builder()
+                        .inputAmount(BigInteger.ONE).outputAmount(BigInteger.ONE).build())
+                    .liquidityPoint(LiquidityPoint.Builder.builder()
+                        .inputAmount(BigInteger.valueOf(5)).outputAmount(BigInteger.TEN).build())
                     .build())
             .appliesTo(InterledgerAddress.of("test1.foo"))
             .sourceHoldDuration(Duration.of(10, ChronoUnit.MINUTES))
